Refactored out template class sk::rt::Callable.
This class now encapsulates a function reference together with parameters (up
to 3) and the knowledge how to invoke them. Implemented Functor and Method it
terms of sk::rt::Callable. It will be used also in sk::rt::Lock and
sk::rt::ConditionMediator. In general, it can be used everywhere where you
need a postponed method/functor invocation.

git-svn-id: https://svn.code.sf.net/p/stemkit/code@2443 faf13d3b-af0d-0410-97ba-fcc57a00092e